fastapi-sheets-api/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
python-dotenv
pydantic
pydantic-settings
python-multipart
email-validator
